+++ packages/@aws-cdk/aws-stepfunctions-tasks/lib/index.ts
@@ -10,9 +10,9 @@ export * from './sqs/send-to-queue';
 export * from './sqs/send-message';
 export * from './ecs/run-ecs-ec2-task';
 export * from './ecs/run-ecs-fargate-task';
-export * from './sagemaker/sagemaker-task-base-types';
-export * from './sagemaker/sagemaker-train-task';
-export * from './sagemaker/sagemaker-transform-task';
+export * from './sagemaker/base-types';
+export * from './sagemaker/create-training-job';
+export * from './sagemaker/create-transform-job';
 export * from './start-execution';
 export * from './stepfunctions/start-execution';
 export * from './evaluate-expression';
